Extract invalidation notification
